+++ src/types.ts
@@ -125,7 +125,7 @@ interface ExtensionConfigType {
 			code?: string;
 		};
 		severity: 'error' | 'warning' | 'info' | 'hint';
-		// decoration: DecorationRenderOptions;
+		decoration: DecorationRenderOptions;
 	}>;
 	/**
 	 * Exclude diagnostics by message. Excludes when the diagnostic message contains the exlcuding string(case-insensitive). Can use Regular Expressions (use object instead of a string).